Refactor Vault token management to use simple token refresher approach
